pyproject.toml
[project]
name = "claimchecker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "google-genai>=1.0.0",
    "langchain-text-splitters>=0.3.0",
    "mlx-lm>=0.31.2",
    "nltk>=3.9.4",
    "pydantic>=2.12.5",
    "pypdf>=5.0.0",
    "python-multipart>=0.0.24",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "torch>=2.11.0",
    "transformers>=5.5.1",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.44.0",
]

[tool.poe.tasks.backend]
cmd = "uvicorn main:app --reload"
cwd = "backend"

[tool.poe.tasks.frontend]
cmd = "streamlit run frontend/app.py"

[tool.poe.tasks.dev]
parallel = ["backend", "frontend"]
